third_party/blink/web_tests/webexposed/null-attributes-shared-worker.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<script>
description("This test interface attributes which should be null in SharedWorkerGlobalScope.");
worker = startWorker("resources/null-attributes-shared-worker.js", "shared");
worker.port.postMessage({});
</script>
