third_party/blink/web_tests/webexposed/null-attributes-shared-worker-expected.txt

This test interface attributes which should be null in SharedWorkerGlobalScope.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/null-attributes-shared-worker.js
PASS [Worker] navigator.serial is null
PASS [Worker] navigator.usb is null
PASS [Worker] navigator.wakeLock is null
PASS successfullyParsed is true

TEST COMPLETE
